+++ src/Holding/ViewHolding/HoldingReceivingHistory/HoldingReceivingHistory.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import user from '@testing-library/user-event';
 import { screen } from '@testing-library/react';
 
 import '../../../../test/jest/__mock__';
@@ -10,6 +11,7 @@ import {
 } from './fixtures';
 import HoldingReceivingHistory from './HoldingReceivingHistory';
 import useReceivingHistory from './useReceivingHistory';
+import { SORT_DIRECTION } from '../../../constants';
 
 jest.mock('./useReceivingHistory', () => jest.fn());
 
@@ -26,5 +28,18 @@ describe('HoldingReceivingHistory', () => {
     renderHoldingReceivingHistory({ id: 'holdingUid' });
 
     expect(screen.getByText(receivingHistory[0].enumeration)).toBeInTheDocument();
+  });
+
+  it('should apply sort by a column', () => {
+    renderHoldingReceivingHistory({ id: 'holdingUid' });
+
+    const captionHeader = screen.getAllByRole('columnheader')[0];
+    const btn = screen.getByRole('button', { name: 'ui-inventory.caption' });
+
+    user.click(btn);
+    expect(captionHeader.getAttribute('aria-sort')).toBe(SORT_DIRECTION.ASCENDING);
+
+    user.click(btn);
+    expect(captionHeader.getAttribute('aria-sort')).toBe(SORT_DIRECTION.DESCENDING);
   });
 });