+++ src/routes/QuickMarcRoute.js
@@ -6,9 +6,12 @@ import { Pluggable } from '@folio/stripes/core';
 
 const QuickMarcRoute = ({ match, history, location }) => {
   const onClose = useCallback((recordRoute) => {
+    const newSearchParams = new URLSearchParams(location.search);
+    newSearchParams.delete('relatedRecordVersion');
+
     history.push({
       pathname: `/inventory/view/${recordRoute}`,
-      search: location.search,
+      search: newSearchParams.toString(),
     });
   }, [location.search]);